+++ releasenotes.md
@@ -9,5 +9,6 @@ Whats New in v12.9
 - Removed time from created and updated fields
 - Removed created and updated timestamps from ingredient suppliers, bottles, customers and formula documents
 - Removed updated timestamp from html templates
-- Fixed CSV formula import
+- Overall improve CSV formula import
+- Auto assign columns to CSV formula import
 - This release may include additional changes; for full details, please refer to the CHANGELOG